refactor: 💡 [api] move middleware.go util funcs
